+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This file follows the best practices from [keepachangelog.com](http://keepachang
 - Added rack-attack for request throttling [#3078](https://github.com/sharetribe/sharetribe/pull/3078)
 - Stripe integration [#3018](https://github.com/sharetribe/sharetribe/pull/3018)
 - Sending emails from admin to specified subset of users [#3058](https://github.com/sharetribe/sharetribe/pull/3058)
+- Custom Scripts are now also enabled in Custom Landing Page [#3080](https://github.com/sharetribe/sharetribe/pull/3080/files)
 
 ### Changed